Fix memory_unified_search FLOAT4/f64 decode + fresh-DB migration ordering
memory_unified_search 500d with "Rust type f64 (FLOAT8) is not compatible
with SQL type FLOAT4": the memory_unified_nodes matview's `importance`
column resolves to `real` (its UNION ALL leads with memory_entities.importance
REAL), but UnifiedNodeHit.importance was typed f64, so SQLx aborts decoding on
the first returned row. Fix: f64 -> f32, matching the matview column and the
five sibling importance structs. Regression from the a65b12b god-file split;
latent until the matview held embedded rows (decode errors fire only on >=1
returned row). Hardened the existing decode-exercising test with an explicit
importance assertion.

Also fix a pre-existing fresh-DB migration ordering bug surfaced while
verifying: ensure_work_item_experiment_bridge ran before the numbered tracker
step creates work_items, so its to_regclass guard silently skipped table
creation and ensure_memory_unified_views then aborted on a fresh DB with
'relation "work_item_experiment" does not exist'. Moved the bridge call to
just before the unified-views build (after work_items exists); corrected the
comments (incl. v6_unified_graph) that rationalized the old ordering.
Idempotent; safe for existing DBs.

Verified: verify.sh all 8 gates green; decode proven at runtime against the
live 611k-row matview via the fixed binary in-process and through the
restarted daemon over MCP; migration fix proven via a fresh template build.

+++ src/db/queries/memory_search.rs
@@ -783,7 +783,12 @@ pub struct UnifiedNodeHit {
     pub node_id: String,
     pub node_type: String,
     pub label: String,
-    pub importance: f64,
+    // `real` (FLOAT4) in the `memory_unified_nodes` matview: the leading UNION
+    // arm selects `memory_entities.importance` (REAL), and PostgreSQL resolves
+    // the column against the trailing numeric literals to `real`. Decoding it as
+    // `f64` aborts at fetch with a FLOAT4/FLOAT8 mismatch. Matches every sibling
+    // importance struct (EntitySearchHit, MemorySemanticHit, EntityRow, …).
+    pub importance: f32,
     pub similarity: Option<f64>,
 }
